+++ Engine/Platform/include/Pyramid/Platform/Windows/Win32OpenGLWindow.hpp
@@ -41,6 +41,7 @@ namespace Pyramid
         int m_width;
         int m_height;
         bool m_shouldClose;
+        WindowResizeState m_resizeState;
     };
 
 } // namespace Pyramid